New 1997 KSK kit
At least 3 pieces of gear are not realistic but I had to make those choices to make the kit possible. 
The helmet is a Russian K-3 that somewhat looks like the one used in 1997. The shotgun is not the real one since I could not ID it from very old low res images so I took one that looked the most similar. And the backpack was added for gameplay reasons.