v1.3.0
- changed/added UKR_2023_MechInf_MULTICAM.sqf
- changed/added GER_2023_KSK_MULTITARN.sqf
- changed/added V5_GER_2023_Fallschirmjäger_Tropentarn.sqf
- changed/added V5_GER_2023_Fallschirmjäger_Flecktarn.sqf
- changed/added GER_ca2020_KSK_Tropentarn.sqf
- changed/added V5_MEC_2023_3CD.sqf
- renamed main loadout folder
- added Full Mission Sets folder
- added AI Presets folder
- updated readme file

All additions of loadouts are due to the renaming of the main folder as it was outdated. 90% of them have not been changed or updated to V5 and might not work. It is expected that loadouts without the V5 tag are tested before being used in a Cluster Community mission.